lunaconWeb-Project/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
blessed==1.17.5
botocore==1.15.49
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.0
Django==3.0.5
django-crispy-forms==1.9.0
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
filelock==3.0.12
future==0.16.0
idna==2.7
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
pathspec==0.5.9
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.0
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
toml==0.10.0
typed-ast==1.4.1
urllib3==1.24.3
virtualenv==20.0.19
wcwidth==0.1.9
websocket-client==0.57.0
wrapt==1.12.1
zipp==3.1.0
